services/notification-service/requirements.txt
pika
fastapi
uvicorn
psycopg2-binary
json
